+++ familyhelper-webapi-stack/src/main/java/com/dwarfeng/familyhelper/webapi/stack/service/life/ActivityTemplateResponseService.java
@@ -1,0 +1,50 @@
+package com.dwarfeng.familyhelper.webapi.stack.service.life;
+
+import com.dwarfeng.familyhelper.life.stack.bean.dto.ActivityTemplateCreateInfo;
+import com.dwarfeng.familyhelper.life.stack.bean.dto.ActivityTemplatePermissionRemoveInfo;
+import com.dwarfeng.familyhelper.life.stack.bean.dto.ActivityTemplatePermissionUpsertInfo;
+import com.dwarfeng.familyhelper.life.stack.bean.dto.ActivityTemplateUpdateInfo;
+import com.dwarfeng.familyhelper.life.stack.bean.entity.ActivityTemplate;
+import com.dwarfeng.familyhelper.webapi.stack.bean.disp.life.DispActivityTemplate;
+import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
+import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.ServiceException;
+import com.dwarfeng.subgrade.stack.service.Service;
+
+/**
+ * 活动模板响应服务。
+ *
+ * @author DwArFeng
+ * @since 1.0.10
+ */
+public interface ActivityTemplateResponseService extends Service {
+
+    boolean exists(LongIdKey key) throws ServiceException;
+
+    ActivityTemplate get(LongIdKey key) throws ServiceException;
+
+    PagedData<ActivityTemplate> all(PagingInfo pagingInfo) throws ServiceException;
+
+    DispActivityTemplate getDisp(LongIdKey key, StringIdKey inspectAccountKey) throws ServiceException;
+
+    PagedData<DispActivityTemplate> allPermittedDisp(StringIdKey accountKey, PagingInfo pagingInfo)
+            throws ServiceException;
+
+    PagedData<DispActivityTemplate> allOwnedDisp(StringIdKey accountKey, PagingInfo pagingInfo) throws ServiceException;
+
+    LongIdKey createActivityTemplate(StringIdKey userKey, ActivityTemplateCreateInfo activityTemplateCreateInfo)
+            throws ServiceException;
+
+    void updateActivityTemplate(StringIdKey userKey, ActivityTemplateUpdateInfo activityTemplateUpdateInfo)
+            throws ServiceException;
+
+    void removeActivityTemplate(StringIdKey userKey, LongIdKey activityTemplateKey) throws ServiceException;
+
+    void upsertPermission(StringIdKey userKey, ActivityTemplatePermissionUpsertInfo permissionUpsertInfo)
+            throws ServiceException;
+
+    void removePermission(StringIdKey userKey, ActivityTemplatePermissionRemoveInfo permissionRemoveInfo)
+            throws ServiceException;
+}